Moved charge efficiencies to the baseclass
