+++ h2integrate/storage/battery/pysam_battery.py
@@ -4,7 +4,7 @@ import PySAM.BatteryStateful as BatteryStateful
 from attrs import field, define
 
 from h2integrate.core.utilities import merge_shared_inputs
-from h2integrate.core.validators import gt_zero, contains, range_val, range_val_or_none
+from h2integrate.core.validators import gt_zero, contains, range_val
 from h2integrate.storage.storage_baseclass import (
     StoragePerformanceBase,
     StoragePerformanceBaseConfig,
@@ -49,19 +49,6 @@ class PySAMBatteryPerformanceModelConfig(StoragePerformanceBaseConfig):
             environment [W/m2*K]. Defaults to 20.
         resistance (int | float, optional): Battery internal resistance [Ohm].
             Defaults to 0.001.
-        charge_efficiency (float | None, optional): Efficiency of charging the battery,
-            represented as a decimal between 0 and 1 (e.g., 0.9 for 90% efficiency).
-            This is applied on top of PySAM's internal chemistry-based efficiency.
-            Optional if ``round_trip_efficiency`` is provided. Defaults to None.
-        discharge_efficiency (float | None, optional): Efficiency of discharging the battery,
-            represented as a decimal between 0 and 1 (e.g., 0.9 for 90% efficiency).
-            This is applied on top of PySAM's internal chemistry-based efficiency.
-            Optional if ``round_trip_efficiency`` is provided. Defaults to None.
-        round_trip_efficiency (float | None, optional): Combined efficiency of charging and
-            discharging the battery, represented as a decimal between 0 and 1 (e.g., 0.81 for
-            81% efficiency). If provided, ``charge_efficiency`` and ``discharge_efficiency``
-            are each set to the square root of this value. Optional if
-            ``charge_efficiency`` and ``discharge_efficiency`` are provided. Defaults to None.
     """
 
     max_capacity: float = field(validator=gt_zero)
@@ -80,30 +67,6 @@ class PySAMBatteryPerformanceModelConfig(StoragePerformanceBaseConfig):
     Cp: int | float = field(default=900)
     battery_h: int | float = field(default=20)
     resistance: int | float = field(default=0.001)
-
-    charge_efficiency: float | None = field(default=None, validator=range_val_or_none(0, 1))
-    discharge_efficiency: float | None = field(default=None, validator=range_val_or_none(0, 1))
-    round_trip_efficiency: float | None = field(default=None, validator=range_val_or_none(0, 1))
-
-    def __attrs_post_init__(self):
-        """Post-initialization to resolve efficiencies.
-
-        If ``round_trip_efficiency`` is provided and individual efficiencies are not,
-        calculates ``charge_efficiency`` and ``discharge_efficiency`` as the square
-        root of ``round_trip_efficiency``. If no efficiency is specified, defaults
-        both to 1.0 (no additional external efficiency loss beyond PySAM's
-        internal chemistry-based efficiency).
-        """
-        if self.round_trip_efficiency is not None and (
-            self.charge_efficiency is None and self.discharge_efficiency is None
-        ):
-            self.charge_efficiency = np.sqrt(self.round_trip_efficiency)
-            self.discharge_efficiency = np.sqrt(self.round_trip_efficiency)
-
-        if self.charge_efficiency is None:
-            self.charge_efficiency = 1.0
-        if self.discharge_efficiency is None:
-            self.discharge_efficiency = 1.0
 
 
 class PySAMBatteryPerformanceModel(StoragePerformanceBase):
